pyproject.toml
[project]
name = "ksmr"
version = "0.1.4"
description = "KsmR: An Ultra-Lightweight Personal AI Assistant"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "KsmR contributors"}
]
keywords = ["ai", "agent", "chatbot", "ksmr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "websockets>=12.0",
    "websocket-client>=1.6.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "readability-lxml>=0.8.0",
    "rich>=13.0.0",
    "croniter>=2.0.0",
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
ksmr = "ksmr.cli.commands:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ksmr"]

[tool.hatch.build.targets.wheel.sources]
"ksmr" = "ksmr"

# Include non-Python files in skills
[tool.hatch.build]
include = [
    "ksmr/**/*.py",
    "ksmr/skills/**/*.md",
    "ksmr/skills/**/*.sh",
]

[tool.hatch.build.targets.sdist]
include = [
    "ksmr/",
    "bridge/",
    "README.md",
    "LICENSE",
]

[tool.hatch.build.targets.wheel.force-include]
"bridge" = "ksmr/bridge"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
